refactor: extract camera-ready reporting helpers
